[DF-5671] Fixed cast warnings
Former-commit-id: 54c8ac263fe4cf0c186d3cbaeea409ccfe3a7e4f

+++ Sources/Internal/Scene3D/Entity.cpp
@@ -51,7 +51,7 @@
 #include "Utils/Random.h"
 #include "Scene3D/Components/ComponentHelpers.h"
 
-#define USE_VECTOR(x) (((1 << x) & vectorComponentsMask) != 0)
+#define USE_VECTOR(x) ((((uint64)1 << x) & vectorComponentsMask) != (uint64)0)
 
 namespace DAVA
 {
